requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
typing-extensions>=4.9.0

# UI
streamlit>=1.31.0
streamlit-chat>=0.1.1

# Vector Database
qdrant-client>=1.14.2

# OpenAI
openai>=1.12.0

# Markdown Processing
markdown-it-py>=3.0.0
python-frontmatter>=1.0.0

# File System
watchdog>=3.0.0

# Utilities
python-multipart>=0.0.6
tenacity>=8.2.3
numpy>=1.26.3  # Required for vector operations

# File extraction
pdfplumber>=0.10.0
python-docx>=1.0.0
openpyxl>=3.0.0
beautifulsoup4>=4.12.0
pandas>=2.0.0 